Refactor test modules into own classes
